backend/requirements.txt
beautifulsoup4==4.12.0
diffusers==0.27.2
Flask==3.0.3
Flask_Cors==4.0.1
numpy==1.23.5
openai==0.27.8
Pillow==9.5.0
Pillow==10.3.0
pydantic==1.10.9
selenium==4.21.0
torch==2.0.1
